backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0
pgvector==0.2.5
psycopg2-binary==2.9.9
openai==1.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
structlog==24.1.0
orjson==3.10.1
redis==5.0.4
rq==1.16.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml-html-clean==0.1.1
